EDU-422 feat: RAG 자동화 기능 추가
